fix bug and pull master

+++ services/engine/gstreamer/avcodec/format_processor/processor_vdec_impl.cpp
@@ -59,8 +59,8 @@ int32_t ProcessorVdecImpl::ProcessOptional(const Format &format)
 
 std::shared_ptr<ProcessorConfig> ProcessorVdecImpl::GetInputPortConfig()
 {
-    CHECK_AND_RETURN_RET(width_ > 0 && width_ < MAX_WIDTH, nullptr);
-    CHECK_AND_RETURN_RET(height_ > 0 && height_ < MAX_HEIGHT, nullptr);
+    CHECK_AND_RETURN_RET(width_ > 0 && static_cast<uint32_t>(width_) < MAX_WIDTH, nullptr);
+    CHECK_AND_RETURN_RET(height_ > 0 && static_cast<uint32_t>(height_) < MAX_HEIGHT, nullptr);
 
     GstCaps *caps = nullptr;
     switch (codecName_) {
@@ -105,7 +105,8 @@ std::shared_ptr<ProcessorConfig> ProcessorVdecImpl::GetInputPortConfig()
 
     config->needCodecData_ = (codecName_ == CODEC_MIMIE_TYPE_VIDEO_AVC && isSoftWare_);
     if (maxInputSize_ > 0) {
-        config->bufferSize_ = (maxInputSize_ > MAX_SIZE) ? MAX_SIZE : static_cast<uint32_t>(maxInputSize_);
+        config->bufferSize_ = (static_cast<uint32_t>(maxInputSize_) > MAX_SIZE) ?
+                              MAX_SIZE : static_cast<uint32_t>(maxInputSize_);
     } else {
         // Memory is aligned to 16 bytes
         constexpr uint32_t alignment = 16;
@@ -117,8 +118,8 @@ std::shared_ptr<ProcessorConfig> ProcessorVdecImpl::GetInputPortConfig()
 
 std::shared_ptr<ProcessorConfig> ProcessorVdecImpl::GetOutputPortConfig()
 {
-    CHECK_AND_RETURN_RET(width_ > 0 && width_ < MAX_WIDTH, nullptr);
-    CHECK_AND_RETURN_RET(height_ > 0 && height_ < MAX_HEIGHT, nullptr);
+    CHECK_AND_RETURN_RET(width_ > 0 && static_cast<uint32_t>(width_) < MAX_WIDTH, nullptr);
+    CHECK_AND_RETURN_RET(height_ > 0 && static_cast<uint32_t>(height_) < MAX_HEIGHT, nullptr);
 
     GstCaps *caps = gst_caps_new_simple("video/x-raw",
         "format", G_TYPE_STRING, gstPixelFormat_.c_str(), nullptr);
